+++ src/Hledger/Flow/Reports.hs
@@ -6,11 +6,14 @@ module Hledger.Flow.Reports
 
 import Turtle hiding (stdout, stderr, proc)
 import Prelude hiding (FilePath, putStrLn, writeFile)
-import qualified Data.Text as T
-import qualified Hledger.Flow.Types as FlowTypes
 import Hledger.Flow.Report.Types
 import Hledger.Flow.Common
 import Control.Concurrent.STM
+import Data.Either
+
+import qualified Data.Text as T
+import qualified Hledger.Flow.Types as FlowTypes
+import qualified Data.List as List
 
 generateReports :: ReportOptions -> IO ()
 generateReports opts = sh (
@@ -19,50 +22,72 @@ generateReports opts = sh (
     logHandle <- fork $ consoleChannelLoop ch
     liftIO $ if (showOptions opts) then channelOutLn ch (repr opts) else return ()
     (reports, diff) <- time $ liftIO $ generateReports' opts ch
-    liftIO $ channelOutLn ch $ format ("Generated "%d%" reports in "%s) (length reports) $ repr diff
+    let failedAttempts = lefts reports
+    let failedText = if List.null failedAttempts then "" else format ("(and attempted to write "%d%" more) ") $ length failedAttempts
+    liftIO $ channelOutLn ch $ format ("Generated "%d%" reports "%s%"in "%s) (length (rights reports)) failedText $ repr diff
     liftIO $ terminateChannelLoop ch
     wait logHandle
   )
 
-generateReports' :: ReportOptions -> TChan FlowTypes.LogMessage -> IO [FilePath]
+generateReports' :: ReportOptions -> TChan FlowTypes.LogMessage -> IO [Either FilePath FilePath]
 generateReports' opts ch = do
-  channelOutLn ch "Report generation has not been fully implemented yet. Keep an eye out for report pull requests: https://github.com/apauley/hledger-flow/pulls"
-  ownerReports opts ch "everyone"
+  let wipMsg = "Report generation is still a work-in-progress - please let me know how this can be more useful.\n\n"
+               <> "Keep an eye out for report-related pull requests and issues, and feel free to submit some of your own:\n"
+               <> "https://github.com/apauley/hledger-flow/pulls\n"
+               <> "https://github.com/apauley/hledger-flow/issues\n"
+  channelOutLn ch wipMsg
+  owners <- single $ shellToList $ listOwners opts
+  let baseJournal = journalFile opts []
+  let baseReportDir = outputDir opts []
+  years <- includeYears ch baseJournal
+  let reportParams = [(baseJournal, baseReportDir)] ++ map (ownerParams opts) owners
+  let actions = List.concat $ fmap (generateReports'' opts ch years) reportParams
+  if (sequential opts) then sequence actions else single $ shellToList $ parallel actions
 
-ownerReports :: ReportOptions -> TChan FlowTypes.LogMessage -> Text -> IO [FilePath]
-ownerReports opts ch owner = do
-  let journal = (baseDir opts) </> "all-years" <.> "journal"
-  let reportsDir = (baseDir opts) </> "reports" </> fromText owner
-  let actions = map (\r -> r opts ch journal reportsDir) [accountList, incomeStatement]
-  results <- if (sequential opts) then sequence actions else single $ shellToList $ parallel actions
-  return $ map fst results
+generateReports'' :: ReportOptions -> TChan FlowTypes.LogMessage -> [Integer] -> (FilePath, FilePath) -> [IO (Either FilePath FilePath)]
+generateReports'' opts ch years (journal, reportsDir) = do
+  y <- years
+  let actions = map (\r -> r opts ch journal reportsDir y) [accountList, incomeStatement]
+  map (fmap fst) actions
 
-incomeStatement :: ReportOptions -> TChan FlowTypes.LogMessage -> FilePath -> FilePath -> IO (FilePath, FlowTypes.FullTimedOutput)
-incomeStatement opts ch journal reportsDir = do
-  mktree reportsDir
-  let outputFile = reportsDir </> "income-expenses" <.> "txt"
+incomeStatement :: ReportOptions -> TChan FlowTypes.LogMessage -> FilePath -> FilePath -> Integer -> IO (Either FilePath FilePath, FlowTypes.FullTimedOutput)
+incomeStatement opts ch journal reportsDir year = do
   let sharedOptions = ["--depth", "2", "--pretty-tables", "not:equity"]
-  let reportArgs = ["incomestatement"] ++ sharedOptions ++ ["--average", "--yearly"]
-  generateReport' opts ch journal outputFile reportArgs
+  let reportArgs = ["incomestatement"] ++ sharedOptions
+  generateReport opts ch journal reportsDir year ("income-expenses" <.> "txt") reportArgs
 
-accountList :: ReportOptions -> TChan FlowTypes.LogMessage -> FilePath -> FilePath -> IO (FilePath, FlowTypes.FullTimedOutput)
-accountList opts ch journal reportsDir = do
-  let outputFile = reportsDir </> "accounts" <.> "txt"
+accountList :: ReportOptions -> TChan FlowTypes.LogMessage -> FilePath -> FilePath -> Integer -> IO (Either FilePath FilePath, FlowTypes.FullTimedOutput)
+accountList opts ch journal reportsDir year = do
   let reportArgs = ["accounts"]
-  generateReport' opts ch journal outputFile reportArgs
+  generateReport opts ch journal reportsDir year ("accounts" <.> "txt") reportArgs
 
-generateReport' :: ReportOptions -> TChan FlowTypes.LogMessage -> FilePath -> FilePath -> [Text] -> IO (FilePath, FlowTypes.FullTimedOutput)
-generateReport' opts ch journal outputFile args = do
-  let reportsDir = directory outputFile
+generateReport :: ReportOptions -> TChan FlowTypes.LogMessage -> FilePath -> FilePath -> Integer -> FilePath -> [Text] -> IO (Either FilePath FilePath, FlowTypes.FullTimedOutput)
+generateReport opts ch journal baseOutDir year fileName args = do
+  let reportsDir = baseOutDir </> intPath year
   mktree reportsDir
+  let outputFile = reportsDir </> fileName
   let relativeJournal = relativeToBase opts journal
-  let reportArgs = ["--file", format fp journal] ++ args
-  let reportDisplayArgs = ["--file", format fp relativeJournal] ++ args
+  let reportArgs = ["--file", format fp journal, "--period", repr year] ++ args
+  let reportDisplayArgs = ["--file", format fp relativeJournal, "--period", repr year] ++ args
   let hledger = format fp $ FlowTypes.hlPath . hledgerInfo $ opts :: Text
   let cmdLabel = format ("hledger "%s) $ showCmdArgs reportDisplayArgs
   result@((exitCode, stdOut, _), _) <- timeAndExitOnErr opts ch cmdLabel dummyLogger channelErr procStrictWithErr (hledger, reportArgs, empty)
-  if not (T.null stdOut) then do
-    writeTextFile outputFile (cmdLabel <> "\n\n"<> stdOut)
-    channelOutLn ch $ format ("Wrote "%fp) $ relativeToBase opts outputFile
-    else channelErrLn ch $ format ("No report output for '"%s%"' "%s) cmdLabel (repr exitCode)
-  return (outputFile, result)
+  if not (T.null stdOut)
+    then
+    do
+      writeTextFile outputFile (cmdLabel <> "\n\n"<> stdOut)
+      channelOutLn ch $ format ("Wrote "%fp) $ relativeToBase opts outputFile
+      return (Right outputFile, result)
+    else
+    do
+      channelErrLn ch $ format ("No report output for '"%s%"' "%s) cmdLabel (repr exitCode)
+      return (Left outputFile, result)
+
+journalFile :: ReportOptions -> [FilePath] -> FilePath
+journalFile opts dirs = (foldl (</>) (baseDir opts) dirs) </> "all-years" <.> "journal"
+
+outputDir :: ReportOptions -> [FilePath] -> FilePath
+outputDir opts dirs = foldl (</>) (baseDir opts) ("reports":dirs)
+
+ownerParams :: ReportOptions -> FilePath -> (FilePath, FilePath)
+ownerParams opts owner = (journalFile opts ["import", owner], outputDir opts [owner])